requirements.txt
ipykernel
streamlit
streamlit-folium
folium
docx2txt
PyPDF2
wordcloud
nltk
numpy
scipy
pandas
geopandas
matplotlib
plotly
neo4j
networkx
tiktoken
wordcloud
twilio
jupyter
yfiles_jupyter_graphs
graphviz
python-dotenv
json-repair
langchain
langchain-community
langchain-ollama              
langchain-openai
langchain-anthropic
langchain-neo4j
langchain_core
langchain-experimental
unstructured>=0.16.12 
onnx>=1.17.0
protobuf==5.29.2
pydantic==2.10.4